More C++ test utilities
